parkings_function_gcp/requirements.txt
requests
google-cloud-bigquery==3.30.0
functions-framework==3.*


